Moved Sevens zip Resources and Added Dialogue Documents
I'm adding the dialogue documents.
If you check, I didn't finish the dialogue for Peasant, Smith and
Dragon. I'm leaving Peasant dialogue open just so we can add funny
things for him to say later. I forgot to add the dialogue for when you
give back the items which isn't too bad. I'm leaving the Dragon dialogue
for when we do Dragon sprint