Altered Dialogue Menu, Fixed Movement, Changed Project Folder Name (Final)
I changed the formating of the dialogue so the text is bigger.

I changed the forcemode to velocitychange and changed the drag/mass on
the player's rigidbody to limit its speed.

I changed the name of the project folder one last time because "Unity
Project" was too vague. Now it's "Curse of Cubes Unity Project".