Goblins now drop health potions.*****
****Bug: It appears that goblins drop 2 health potions for some reason.

What I changed:
In the enemy file, there is a public GameObject variable in named
health.

I dragged the health potion prefab into the reference for health in the
Unity inspector.

When the goblin dies, it calls the Instantiate function to create the
health potion object.

But when I pick up the potion, the number of health potions in your
inventory increases by 2. ??