Quest items Removeable from Inventory. UsemItem.
Made wand removeable from inventory, as well as set the base for
removing the rest of the quest items.

I changed UseItem to UsePotion for healing. and added RemoveWand using
the same code to remove the wand. I'm going to make a function to remove
each item. It's technically not the best programming practice, but it
works and thats what matters ;p.

Also forgot to put enemyhealth on tutorial goblin and put it back on
him.