Fixed Camera Jittering
The camera was jittering due to whenever the player collided with
something it would stutter in the Y axis. So I freezed the Y axis for
the players rotation, then completely froze the camera so it can't move
from what it's supposed to do.

I also changed the box collider to a sphere collider for the player so
they won't get stuck on edges.