Added Game Over Screen
Added game over screen when player dies. This goes back to the main
menu. I also have code set up for when we add in the multiple endings.
(Dovahkube and Dovahkiid endings)

Also deleted the demo folder in the pyroparticles folder because it was
preventing the project from being Built and Ran.