Changed Dragon Attack Pattern
I had to first change the their box collider sizes and positions because
the fireballs couldn't get through it because it was a full collider.

Then I noticed that for some reason, when you collided with the dragons
box collider you couldn't get hurt by the fireballs. Making it trigger
allowed you not to hide from the fireballs but that means no collision
at all with dragon. So I made the dialoguebox collider they have and
made it to scale. Then I noticed that colliding with the trigger
collider now makes them go aggro instantly, so I made the dialoguebox
collider infront of the regular box collider so that wouldn't happen.

I think I could probably change it back to collision and just have the
dialoguebox out in front but I can test that after.